+++ createRollupConfig.js
@@ -2,7 +2,9 @@ import { terser } from 'rollup-plugin-terser';
 import peerDepsExternal from 'rollup-plugin-peer-deps-external';
 import { getBabelOutputPlugin } from '@rollup/plugin-babel';
 
-const babelConfig = {
+const { name } = require(__dirname + '/package.json');
+
+const babelBaseConfig = {
   babelrc: false,
   presets: [
     [
@@ -15,6 +17,23 @@ const babelConfig = {
     ],
   ],
 };
+
+const babelUMDConfig = {
+  babelrc: false,
+  allowAllFormats: true,
+  presets: [
+    [
+      '@babel/env',
+      {
+        modules: false,
+      },
+    ],
+  ],
+};
+
+function getGlobalName(name) {
+  return `__${name.replace(/[@\-/]/g, '')}__`;
+}
 
 function createRollupConfig() {
   const config = {
@@ -50,25 +69,40 @@ function createRollupConfig() {
         file: `./lib/index.es5.js`,
         format: 'cjs',
         exports: 'named',
-        plugins: [getBabelOutputPlugin(babelConfig)],
+        plugins: [getBabelOutputPlugin(babelBaseConfig)],
       },
       {
         file: `./lib/index.es5.min.js`,
         format: 'cjs',
         exports: 'named',
-        plugins: [getBabelOutputPlugin(babelConfig), terser()],
+        plugins: [getBabelOutputPlugin(babelBaseConfig), terser()],
       },
       {
         file: `./lib/index.es5.mjs`,
         format: 'esm',
         exports: 'named',
-        plugins: [getBabelOutputPlugin(babelConfig)],
+        plugins: [getBabelOutputPlugin(babelBaseConfig)],
       },
       {
         file: `./lib/index.es5.min.mjs`,
         format: 'esm',
         exports: 'named',
-        plugins: [getBabelOutputPlugin(babelConfig), terser()],
+        plugins: [getBabelOutputPlugin(babelBaseConfig), terser()],
+      },
+      {
+        file: `./lib/index.umd.js`,
+        format: 'umd',
+        name: getGlobalName(name),
+        plugins: [getBabelOutputPlugin(babelUMDConfig), terser()],
+        globals: {
+          '@merkur/core': getGlobalName('@merkur/core'),
+          '@merkur/plugin-component': getGlobalName('@merkur/plugin-component'),
+          '@merkur/plugin-event': getGlobalName('@merkur/plugin-event'),
+          '@merkur/plugin-http-client': getGlobalName(
+            '@merkur/plugin-http-client'
+          ),
+          'node-fetch': 'fetch',
+        },
       },
     ],
   };